Fix DigitalOcean path; separate experimental stuff
