+++ inngest/experimental/__init__.py
@@ -1,0 +1,5 @@
+"""
+Experimental entrypoint for the Inngest SDK.
+
+Does not follow semantic versioning! Breaking changes may occur at any time.
+"""
