+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "inngest"
-version = "0.3.13.dev0"
+version = "0.3.13.dev1"
 description = "Python SDK for Inngest"
 readme = "README.md"
 classifiers = [
