+++ tests/test_introspection/test_digital_ocean.py
@@ -6,7 +6,8 @@ import flask.testing
 import inngest
 import inngest.digital_ocean
 import inngest.fast_api
-from inngest._internal import const, digital_ocean_simulator
+from inngest._internal import const
+from inngest.experimental import digital_ocean_simulator
 from tests import base
 
 _framework = const.Framework.DIGITAL_OCEAN.value
